requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-celery==3.3.1
django-environ==0.4.5
django-lifecycle==0.9.1
django-redis==4.12.1
django-split-settings==1.0.1
djangorestframework==3.12.2
drf-yasg==1.20.0
Faker==7.0.1
flake8==3.8.4
flake8-builtins==1.5.3
flake8-django==1.1.1
flake8-import-order==0.18.1
flake8-print==4.0.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==3.0.37
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.9
pathlib==1.0.1
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.3
urlman==1.4.0
